modularize the todo state manager
